refactoring e2e elasticsearch test
